// grau-frontend-projekt/src/app/models/traffic.ts
// Generated by https://quicktype.io

export interface Traffic {
  timestamps:       { [key: string]: string };
  RC1?:              { [key: string]: number };
  RC2?:              { [key: string]: number };
  RC3?:              { [key: string]: number };
  RC4?:              { [key: string]: number };
  RC5?:              { [key: string]: number };
  SC1?:              { [key: string]: number };
  SC2?:              { [key: string]: number };
  SC3?:              { [key: string]: number };
  SC4?:              { [key: string]: number };
  SC5?:              { [key: string]: number };
  Motorrad:         { [key: string]: number };
  PKW:              { [key: string]: number };
  Kleintransporter: { [key: string]: number };
  Transporter:      { [key: string]: number };
  LKW:              { [key: string]: number };
}